implement mp3 output format
