+++ kokoros-openai/Cargo.toml
@@ -9,5 +9,3 @@ kokoros = { path = "../kokoros" }
 axum = { version = "0.7", features = ["http2"] }
 serde = { version = "1.0", features = ["derive"] }
 tower-http = { version = "0.5", features = ["cors"] }
-
-mp3lame-encoder = "0.2.1"
