+++ kokoros/Cargo.toml
@@ -14,6 +14,8 @@ reqwest = { version = "0.12.12" }
 serde_json = "1.0.135"
 tokio = { version = "1.0", features = ["fs", "io-util"] }
 ndarray-npy = "0.9.1"
+mp3lame-encoder = "0.2.1"
+
 
 # ONNX Runtime dependencies
 ort = { version = "2.0.0-rc.9", default-features = true }
