+++ kokoros/src/utils/mod.rs
@@ -1,2 +1,3 @@
 pub mod fileio;
 pub mod wav;
+pub mod mp3;
